requirements.txt
# Core dependencies
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Deep learning
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# XAI
pytorch-grad-cam>=1.4.0

# Data processing
pandas>=2.0.0
pyarrow>=12.0.0  # for parquet support

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# LLM
openai>=1.0.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
